feat(react-18-toast): customize createStandaloneToasts like ToastProvider
