+++ packages/toast/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./use-toast"
-export * from "./toast-provider"
+export * from "./toast.provider"
 export * from "./toast.types"
 export type { ToastPosition, ToastPositionWithLogical } from "./toast.placement"
-export { createStandaloneToast } from "./create-standalone-toast"
-export { Toast } from "./toast"
-export { createRenderToast } from "./toast"
+export * from "./create-standalone-toast"
+export * from "./toast"
